move test reports to own file
